requirements.txt
customtkinter==5.0.3
clipspy==1.0.4